Update release-trueos.sh
Cleanup a bit of the basename cleanup, and add some extra checks to the ZFS dataset validation/removal.

+++ release/release-trueos.sh
@@ -166,14 +166,21 @@ super_clean_poudriere()
 	#Look for any leftover mountpoints/directories and remove them too
 	for i in `ls ${POUDRIERED_DIR}/*-make.conf`
 	do
-		name=`basename "${i}" | rev | cut -d "-" -f 2-10 | rev`
+		name=`basename "${i}" | sed 's|-make.conf||g'`
 		if [ ! -d "${POUDRIERED_DIR}/jails/${name}" ]  && [ -d "${POUDRIERE_BASEFS}/jails/${name}" ] ; then
 			#Jail configs missing, but jail mountpoint still exists
 			#Need to completely destroy the old jail dataset/dir
 			_stale_dataset=$(mount | grep 'on ${POUDRIERE_BASEFS}/jails/${name} ' | cut -w -f 1)
-			umount ${POUDRIERE_BASEFS}/jails/${name}
-			rmdir ${POUDRIERE_BASEFS}/jails/${name}
-			zfs destroy -r ${_stale_dataset}
+			if [ -n "${_stale_dataset}" ] ; then
+				#Found a stale mountpoint/dataset
+				umount ${POUDRIERE_BASEFS}/jails/${name}
+				rmdir ${POUDRIERE_BASEFS}/jails/${name}
+				#Verify that it is a valid ZFS dataset
+				zfs list "${_stale_dataset}"
+				if [ 0 -eq $? ] ; then
+					zfs destroy -r ${_stale_dataset}
+				fi
+			fi
 		fi
 	done
 }
